requirements.txt
# ═══════════════════════════════════════════════
#   MediAgent v5 — Python Requirements
# ═══════════════════════════════════════════════

# ── Core ──────────────────────────────────────
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Database ──────────────────────────────────
sqlalchemy==2.0.25

# ── HTTP (async-capable) ──────────────────────
httpx==0.26.0
requests==2.31.0

# ── Security / Rate limiting ──────────────────
slowapi==0.1.9

# ── Background scheduler ──────────────────────
apscheduler==3.10.4

# ── PDF export ────────────────────────────────
reportlab==4.0.9

# ── WhatsApp (Twilio) ─────────────────────────
twilio==8.13.0

# ── Validation ────────────────────────────────
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# ── File type detection (magic bytes) ─────────
# python-magic requires libmagic:
#   Ubuntu/Debian: sudo apt-get install libmagic1
#   macOS:         brew install libmagic
#   Windows:       pip install python-magic-bin
python-magic==0.4.27
